separate rwd components
